Исправлена реализация DiyArrayList.
